refactor: api fetch와 token 검사를 별도의 함수로 분리
