+++ src/api/line.js
@@ -1,23 +1,14 @@
-import { SESSION_STORAGE_KEY, URL } from '../constants.js';
-import { sessionStore } from '../utils/utils.js';
+import { URL } from '../constants.js';
+import { sendGetRequest, sendPostRequest, sendPutRequest, sendDeleteRequest } from '../utils/api.js';
 
 export const requestLineRegistration = async line => {
-  const accessToken = sessionStore.getItem(SESSION_STORAGE_KEY.ACCESS_TOKEN);
-  if (!accessToken) return;
-  const response = await fetch(`${URL.BASE_URL}/lines`, {
-    method: 'POST',
-    headers: {
-      Authorization: `Bearer ${accessToken}`,
-      'Content-Type': 'application/json;charset=UTF-8',
-    },
-    body: JSON.stringify({
-      name: line.name,
-      color: line.color,
-      upStationId: Number(line.upStationId),
-      downStationId: Number(line.downStationId),
-      distance: Number(line.distance),
-      duration: Number(line.duration),
-    }),
+  const response = await sendPostRequest(`${URL.BASE_URL}/lines`, {
+    name: line.name,
+    color: line.color,
+    upStationId: Number(line.upStationId),
+    downStationId: Number(line.downStationId),
+    distance: Number(line.distance),
+    duration: Number(line.duration),
   });
 
   const newLine = await response.json();
@@ -37,16 +28,7 @@ export const requestLineRegistration = async line => {
 };
 
 export const requestLineList = async () => {
-  const accessToken = sessionStore.getItem(SESSION_STORAGE_KEY.ACCESS_TOKEN);
-  if (!accessToken) return;
-  const response = await fetch(`${URL.BASE_URL}/lines`, {
-    method: 'GET',
-    headers: {
-      Authorization: `Bearer ${accessToken}`,
-      'Content-Type': 'application/json;charset=UTF-8',
-    },
-  });
-
+  const response = await sendGetRequest(`${URL.BASE_URL}/lines`);
   const newLineList = await response.json();
 
   if (!response.ok) {
@@ -65,14 +47,7 @@ export const requestLineList = async () => {
 };
 
 export const requestLineDelete = async lineId => {
-  const accessToken = sessionStore.getItem(SESSION_STORAGE_KEY.ACCESS_TOKEN);
-  if (!accessToken) return;
-  const response = await fetch(`${URL.BASE_URL}/lines/${lineId}`, {
-    method: 'DELETE',
-    headers: {
-      Authorization: `Bearer ${accessToken}`,
-    },
-  });
+  const response = await sendDeleteRequest(`${URL.BASE_URL}/lines/${lineId}`);
 
   if (!response.ok) {
     throw new Error('역 등록 실패');
@@ -80,18 +55,9 @@ export const requestLineDelete = async lineId => {
 }
 
 export const requestLineUpdate = async (newline) => {
-  const accessToken = sessionStore.getItem(SESSION_STORAGE_KEY.ACCESS_TOKEN);
-  if (!accessToken) return;
-  const response = await fetch(`${URL.BASE_URL}/lines/${newline.id}`, {
-    method: 'PUT',
-    headers: {
-      Authorization: `Bearer ${accessToken}`,
-      'Content-Type': 'application/json;charset=UTF-8',
-    },
-    body: JSON.stringify({
-      name: newline.name,
-      color: newline.color,
-    }),
+  const response = await sendPutRequest(`${URL.BASE_URL}/lines/${newline.id}`, {
+    name: newline.name,
+    color: newline.color,
   });
 
   if (!response.ok) {
