+++ src/api/station.js
@@ -1,18 +1,8 @@
-import { sessionStore } from '../utils/utils.js';
-import { SESSION_STORAGE_KEY, URL } from '../constants.js';
+import { URL } from '../constants.js';
+import { sendGetRequest, sendPostRequest, sendDeleteRequest, sendPutRequest } from '../utils/api.js';
 
 export const requestStationRegistration = async stationName => {
-  const accessToken = sessionStore.getItem(SESSION_STORAGE_KEY.ACCESS_TOKEN);
-  if (!accessToken) return;
-  const response = await fetch(`${URL.BASE_URL}/stations`, {
-    method: 'POST',
-    headers: {
-      Authorization: `Bearer ${accessToken}`,
-      'Content-Type': 'application/json;charset=UTF-8',
-    },
-    body: JSON.stringify({ name: stationName }),
-  });
-  
+  const response = await sendPostRequest(`${URL.BASE_URL}/stations`, { name: stationName });
   const { id, name } = await response.json();
 
   if (!response.ok) {
@@ -23,14 +13,7 @@ export const requestStationRegistration = async stationName => {
 };
 
 export const requestStationList = async () => {
-  const accessToken = sessionStore.getItem(SESSION_STORAGE_KEY.ACCESS_TOKEN);
-  if (!accessToken) return;
-  const response = await fetch(`${URL.BASE_URL}/stations`, {
-    method: 'GET',
-    headers: {
-      Authorization: `Bearer ${accessToken}`,
-    },
-  });
+  const response = await sendGetRequest(`${URL.BASE_URL}/stations`);
 
   if (!response.ok) {
     throw new Error('역 조회 실패');
@@ -41,14 +24,7 @@ export const requestStationList = async () => {
 };
 
 export const requestStationDelete = async stationId => {
-  const accessToken = sessionStore.getItem(SESSION_STORAGE_KEY.ACCESS_TOKEN);
-  if (!accessToken) return;
-  const response = await fetch(`${URL.BASE_URL}/stations/${stationId}`, {
-    method: 'DELETE',
-    headers: {
-      Authorization: `Bearer ${accessToken}`,
-    },
-  });
+  const response = await sendDeleteRequest(`${URL.BASE_URL}/stations/${stationId}`);
 
   if (!response.ok) {
     throw new Error('역 삭제 실패');
@@ -56,16 +32,7 @@ export const requestStationDelete = async stationId => {
 };
 
 export const requestStationUpdate = async (stationId, stationName) => {
-  const accessToken = sessionStore.getItem(SESSION_STORAGE_KEY.ACCESS_TOKEN);
-  if (!accessToken) return;
-  const response = await fetch(`${URL.BASE_URL}/stations/${stationId}`, {
-    method: 'PUT',
-    headers: {
-      Authorization: `Bearer ${accessToken}`,
-      'Content-Type': 'application/json;charset=UTF-8',
-    },
-    body: JSON.stringify({ name: stationName }),
-  });
+  const response = await sendPutRequest(`${URL.BASE_URL}/stations/${stationId}`, { name: stationName });
 
   if (!response.ok) {
     throw new Error('역 수정 실패');
