+++ src/utils/api.js
@@ -1,0 +1,56 @@
+import { sessionStore } from './utils.js';
+import { SESSION_STORAGE_KEY } from '../constants.js';
+
+export const sendGetRequest = async (url, ) => {
+  const accessToken = sessionStore.getItem(SESSION_STORAGE_KEY.ACCESS_TOKEN);
+  if (!accessToken) return { ok: false };
+
+  return await fetch(url, {
+    method: 'GET',
+    headers: {
+      Authorization: `Bearer ${accessToken}`,
+      'Content-Type': 'application/json;charset=UTF-8',
+    },
+  });
+}
+
+export const sendPostRequest = async (url, body) => {
+  const accessToken = sessionStore.getItem(SESSION_STORAGE_KEY.ACCESS_TOKEN);
+  if (!accessToken) return { ok: false };
+
+  return await fetch(url, {
+    method: 'POST',
+    headers: {
+      Authorization: `Bearer ${accessToken}`,
+      'Content-Type': 'application/json;charset=UTF-8',
+    },
+    body: JSON.stringify(body)
+  });
+}
+
+export const sendPutRequest = async (url, body) => {
+  const accessToken = sessionStore.getItem(SESSION_STORAGE_KEY.ACCESS_TOKEN);
+  if (!accessToken) return { ok: false };
+
+  return await fetch(url, {
+    method: 'PUT',
+    headers: {
+      Authorization: `Bearer ${accessToken}`,
+      'Content-Type': 'application/json;charset=UTF-8',
+    },
+    body: JSON.stringify(body)
+  });
+}
+
+export const sendDeleteRequest = async (url) => {
+  const accessToken = sessionStore.getItem(SESSION_STORAGE_KEY.ACCESS_TOKEN);
+  if (!accessToken) return { ok: false };
+
+  return await fetch(url, {
+    method: 'DELETE',
+    headers: {
+      Authorization: `Bearer ${accessToken}`,
+      'Content-Type': 'application/json;charset=UTF-8',
+    },
+  });
+}
